Add functionality - add rating and updating restaurant average rating

+++ src/main/java/com/att/tdp/bisbis10/rating/RatingDTO.java
@@ -7,14 +7,14 @@ import jakarta.validation.constraints.NotNull;
 public class RatingDTO {
     @NotNull(message = "restaurantId is required")
     @Min(value = 1, message = "restaurantId must be at least 1")
-    private Integer restaurantId;
+    private Long restaurantId;
 
     @NotNull(message = "rating is required")
     @Min(value = 0, message = "rating must be at least 0")
     @Max(value = 5, message = "rating can be a max of 5")
     private Float rating;
 
-    public Integer getRestaurantId() {
+    public Long getRestaurantId() {
         return restaurantId;
     }
